fixed fmm3dbie integer compatibility

+++ fortran/fmm-helper-routs.f90
@@ -15,16 +15,16 @@ subroutine lap_comb_dir_eval_addsub_vec(npatches, norders, ixyzs, &
   !   pot: complex *16 (ndim,npts)
 
   implicit none
-  integer npatches,npts,ndd,ndz,ndi,nnz,ndtarg
-  integer norders(npatches),ixyzs(npatches+1),iptype(npatches),ipars(ndi)
+  integer *8 npatches,npts,ndd,ndz,ndi,nnz,ndtarg
+  integer *8 norders(npatches),ixyzs(npatches+1),iptype(npatches),ipars(ndi)
   real *8 srccoefs(9,npts),srcvals(12,npts),eps,dpars(ndd)
   complex *16 zpars(ndz)
-  integer ntarg,nquad,nker,nptso,lwork,ndim
-  integer row_ptr(ntarg+1),col_ind(nnz),iquad(nnz+1),novers(npatches+1)
+  integer *8 ntarg,nquad,nker,nptso,lwork,ndim
+  integer *8 row_ptr(ntarg+1),col_ind(nnz),iquad(nnz+1),novers(npatches+1)
   complex *16 wnear(nquad),sigma(ndim,npts)
-  integer ixyzso(npatches+1)
+  integer *8 ixyzso(npatches+1)
   real *8 targs(ndtarg,ntarg),srcover(12,nptso),whtsover(nptso),work(lwork)
-  integer idensflag, ipotflag, ndim_p
+  integer *8 idensflag, ipotflag, ndim_p
   complex *16 pot(ndim,ntarg)
 
   real *8 potr(ndim,ntarg),poti(ndim,ntarg)
@@ -67,16 +67,16 @@ subroutine helm_comb_dir_eval_addsub_vec(npatches, norders, ixyzs, &
   !   pot: complex *16 (ndim,npts)
 
   implicit none
-  integer npatches,npts,ndd,ndz,ndi,nnz,ndtarg
-  integer norders(npatches),ixyzs(npatches+1),iptype(npatches),ipars(ndi)
+  integer *8 npatches,npts,ndd,ndz,ndi,nnz,ndtarg
+  integer *8 norders(npatches),ixyzs(npatches+1),iptype(npatches),ipars(ndi)
   real *8 srccoefs(9,npts),srcvals(12,npts),eps,dpars(ndd)
   complex *16 zpars(ndz)
-  integer ntarg,nquad,nker,nptso,lwork,ndim
-  integer row_ptr(ntarg+1),col_ind(nnz),iquad(nnz+1),novers(npatches+1)
+  integer *8 ntarg,nquad,nker,nptso,lwork,ndim
+  integer *8 row_ptr(ntarg+1),col_ind(nnz),iquad(nnz+1),novers(npatches+1)
   complex *16 wnear(nquad),sigma(ndim,npts)
-  integer ixyzso(npatches+1)
+  integer *8 ixyzso(npatches+1)
   real *8 targs(ndtarg,ntarg),srcover(12,nptso),whtsover(nptso),work(lwork)
-  integer idensflag, ipotflag, ndim_p
+  integer *8 idensflag, ipotflag, ndim_p
   complex *16 pot(ndim,ntarg)
 
   integer i
@@ -103,20 +103,20 @@ subroutine mtxbsigma_eval_addsub(npatches,norders,ixyzs, &
      ndim,sigma,rjvec,rho,pot,curlgrad)
   
   implicit none
-  integer npatches,npts,nnz_gc,ndtarg
-  integer norders(npatches),ixyzs(npatches+1)
-  integer iptype(npatches)
-  integer ipars
+  integer *8 npatches,npts,nnz_gc,ndtarg
+  integer *8 norders(npatches),ixyzs(npatches+1)
+  integer *8 iptype(npatches)
+  integer *8 ipars
   real *8 srccoefs(9,npts),srcvals(12,npts),eps(2)
   real *8 dpars
   complex *16 zpars(3)
-  integer ntarg,nquad_gc,nptso_gc
-  integer ndim
+  integer *8 ntarg,nquad_gc,nptso_gc
+  integer *8 ndim
   
-  integer row_ptr_gc(ntarg+1),col_ind_gc(nnz_gc)
-  integer iquad_gc(nnz_gc+1),novers_gc(npatches)
+  integer *8 row_ptr_gc(ntarg+1),col_ind_gc(nnz_gc)
+  integer *8 iquad_gc(nnz_gc+1),novers_gc(npatches)
   complex *16 wnear_gc(nquad_gc,3)
-  integer ixyzso_gc(npatches+1)
+  integer *8 ixyzso_gc(npatches+1)
   real *8 srcover_gc(12,nptso_gc)
   real *8 whtsover_gc(nptso_gc)
   
@@ -126,14 +126,14 @@ subroutine mtxbsigma_eval_addsub(npatches,norders,ixyzs, &
   real *8 work
   complex *16 pot(ndim,ntarg)
 
-  integer ixyzso_helm(npatches+1)
-  integer nnz_helm,row_ptr_helm(ntarg+1)
-  integer col_ind_helm(nnz_helm),nquad_helm
-  integer iquad_helm(nnz_helm+1)
+  integer *8 ixyzso_helm(npatches+1)
+  integer *8 nnz_helm,row_ptr_helm(ntarg+1)
+  integer *8 col_ind_helm(nnz_helm),nquad_helm
+  integer *8 iquad_helm(nnz_helm+1)
   complex *16 rjvec(3,npts),rho(npts)
   complex *16 wnear_helm(nquad_helm)
 
-  integer novers_helm(npatches),nptso_helm
+  integer *8 novers_helm(npatches),nptso_helm
   real *8 srcover_helm(12,nptso_helm),whtsover_helm(nptso_helm)
   complex *16 curlj(3,ntarg),gradrho(3,ntarg),curlgrad(6,ntarg)
